chore: Refractor setup in interest modal to another file
